requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
discord.py==2.2.2
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
pycparser==2.21
pydantic==1.10.7
PyNaCl==1.5.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
yarl==1.8.2
